use c++ std::* math functions This updates usages of sin, cos, tan, pow, exp, log, log10, sqrt, fmod, fabs, and fabsf, excluding any usages that look like they might be part of a submodule or 3rd-party code. There's probably some std math functions not listed here that haven't been updated yet.

+++ include/QuadratureLfo.h
@@ -77,11 +77,10 @@ public:
 
 	void tick( float *l, float *r )
 	{
-		*l = sinf( m_phase );
-		*r = sinf( m_phase + m_offset );
+		*l = std::sin(m_phase);
+		*r = std::sin(m_phase + m_offset);
 		m_phase += m_increment;
-
-		while (m_phase >= numbers::tau)	{	m_phase -= numbers::tau; }
+		while (m_phase >= numbers::tau)	{ m_phase -= numbers::tau; }
 	}
 
 private:
